+++ assets/javascripts/kitten/karl/pages/homepage/carousel/index.js
@@ -5,45 +5,58 @@ import { withKnobs, number, object, boolean } from '@storybook/addon-knobs'
 import { CrowdfundingCard } from '../../../../components/cards/crowdfunding-card'
 import { Carousel } from '../../../../components/carousel/carousel/carousel'
 import { CONTAINER_PADDING } from '../../../../constants/grid-config'
+import { Container } from '../../../../components/grid/container'
+import { mediaQueries } from '../../../../hoc/media-queries'
 
 const data = [
   {
-    title: 'Item A',
+    title: 'Atelier d’optique bruxellois dédié à l’upcycling et au cust…',
     imageSrc: 'https://source.unsplash.com/random/500x300?kitten,0',
     thumbSrc: 'https://source.unsplash.com/random/80x80?kitten,1',
   },
   {
-    title: 'Item B',
+    title: 'Keep It Silky',
     imageSrc: 'https://source.unsplash.com/random/500x300?kitten,2',
     thumbSrc: 'https://source.unsplash.com/random/80x80?kitten,3',
   },
   {
-    title: 'Item C',
+    title: '"Odyssée" une exposition itinérante de Fred Kleinberg encore plus long et tout ça',
     imageSrc: 'https://source.unsplash.com/random/500x300?kitten,4',
     thumbSrc: 'https://source.unsplash.com/random/80x80?kitten,5',
   },
   {
-    title: 'Item D',
+    title: 'Les Arts dessinés : devenez abonné-fondateur',
     imageSrc: 'https://source.unsplash.com/random/500x300?kitten,6',
     thumbSrc: 'https://source.unsplash.com/random/80x80?kitten,7',
   },
   {
-    title: 'Item E',
+    title: 'Cras mattis consectetur purus sit amet fermentum.',
     imageSrc: 'https://source.unsplash.com/random/500x300?kitten,8',
     thumbSrc: 'https://source.unsplash.com/random/80x80?kitten,9',
   },
+  {
+    title: 'Maecenas faucibus mollis interdum.',
+    imageSrc: 'https://source.unsplash.com/random/500x300?kitten,9',
+    thumbSrc: 'https://source.unsplash.com/random/80x80?kitten,8',
+  },
+  {
+    title: 'Morbi leo risus, porta ac consectetur ac, vestibulum at eros. Donec id elit non mi porta gravida at eget metus.',
+    imageSrc: 'https://source.unsplash.com/random/500x300?kitten,7',
+    thumbSrc: 'https://source.unsplash.com/random/80x80?kitten,6',
+  },
 ]
 
-export const HomepageCarousel = () => (
-    <>
+const HomepageCarouselBase = ({ viewportIsMOrLess, viewportIsSOrLess }) => (
+    <Container fullWidthBelowScreenSize="M" >
       <Carousel
-        itemMinWidth={280}
+        itemMinWidth={viewportIsSOrLess ? 150 : 280}
         baseItemMarginBetween={CONTAINER_PADDING}
         paginationPosition={{
-          default: 'bottom',
+          default: 'bottom-right',
         }}
-        hidePaginationOnMobile={true}
-        showOtherPages={true}
+        hidePagination={viewportIsMOrLess ? true : false}
+        showOtherPages={viewportIsMOrLess ? true : false}
+        showPageSquares={viewportIsMOrLess ? false : true}
       >
         {data.map(item => (
           <CrowdfundingCard
@@ -62,7 +75,7 @@ export const HomepageCarousel = () => (
             ownerTitle="Owner title"
             ownerDescription="Owner description"
             titleProps={{
-              tag: 'h4',
+              tag: 'h3',
             }}
             cardTitle={item.title}
             cardSubTitle="Custom subtitle"
@@ -72,5 +85,11 @@ export const HomepageCarousel = () => (
           />
         ))}
       </Carousel>
-    </>
+    </Container>
   )
+
+
+export const HomepageCarousel = mediaQueries(HomepageCarouselBase, {
+  viewportIsMOrLess: true,
+  viewportIsSOrLess: true,
+})