+++ CHANGELOG.md
@@ -10,6 +10,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Feature: Add `RocketIcon` and `RocketCircleIcon` components.
 - Feature: Create `ProgressRing` component.
 - Doc: Add a doc for style `utilities`.
+- Feature: Add pagination position `bottom-left` and `bottom-right` on Carousel.
+- Feature: Add pagination squares via `showPageSquares` prop on Carousel.
 
 ## [2.64.0] - 2020-04-07